% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_brk.R
\docType{data}
\name{brk_capaian_provinsi}
\alias{brk_capaian_provinsi}
\title{Capaian dari Pengadaan Pemerintah Provinsi}
\format{Dataframe dengan 792 baris dan 7 kolom

\describe{
  \item{nama_provinsi}{Nama provinsi}
  \item{tahun}{Tahun}
  \item{bulan}{Bulan (dalam angka)}
  \item{capaian_pengadaan}{Capaian Pengadaan Kementerian Lembaga (dalam} persen)
  \item{capaian_kontrak}{Capaian Kontrak Kementerian Lembaga (dalam} persen)
  \item{capaian_pelaksanaan}{Capaian Pelaksanaan Kementerian Lembaga (dalam persen)}
  \item{capaian_pho}{Capaian Provisional Hand Over Kementerian Lembaga (dalam persen)}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/capaian-dari-pengadaan-pemerintah-provinsi}
}
\usage{
brk_capaian_provinsi
}
\description{
Dataset ini berisi persentase dari capaian dari pengadaan di setiap provinsi di Indonesia
}
\keyword{datasets}
