% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_brk.R
\docType{data}
\name{brk_haji}
\alias{brk_haji}
\title{Daftar Tunggu Haji}
\format{Dataframe dengan 142 baris dan 5 kolom

\describe{
  \item{nama_daerah}{Nama Daerah}
  \item{kuota}{Kuota}
  \item{batas_tahun}{Batas sampai dengan tahun kuota tersebut}
  \item{jumlah_pendaftaran}{Jumlah jemaah yang mendaftar}
  \item{jumlah_lunas_tunda}{Jumlah jemaah yang lunas tunda}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/daftar-tunggu-haji}
}
\usage{
brk_haji
}
\description{
Dataset ini berisi daftar tunggu jemaah haji per daerah (provinsi dan kabupaten)
}
\keyword{datasets}
