% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ddk.R
\docType{data}
\name{ddk_sma_putus_sekolah}
\alias{ddk_sma_putus_sekolah}
\title{Siswa Sekolah Menengah Atas yang Putus Sekolah}
\format{Dataframe denan 132 baris dan 4 kolom

\describe{
  \item{nama_provinsi}{Nama Provinsi}
  \item{tahun}{Tahun}
  \item{jumlah_putus_sekolah}{Jumlah siswa SMP yang putus sekolah}
  \item{persen_putus_sekolah}{Persentase siswa SMP yang putus sekolah }
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/siswa-sekolah-menengah-atas-yang-putus-sekolah}
}
\usage{
ddk_sma_putus_sekolah
}
\description{
Dataset ini berisi mengenai jumlah siswa SMA yang putus sekolah per Provinsi
}
\keyword{datasets}
