% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ddk.R
\docType{data}
\name{ddk_tingkat_pendidikan}
\alias{ddk_tingkat_pendidikan}
\title{Tingkat Pendidikan Anak Miskin}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 462 rows and 4 columns.}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/tingkat-pendidikan-anak-miskin}
}
\usage{
ddk_tingkat_pendidikan
}
\description{
Dataset ini berisi data mengenai jumlah anak dengan status kesejahteraan 30% yang bersekolah berdasarkan tingkat pendidikan dan jenis kelamin
}
\keyword{datasets}
