% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eko.R
\docType{data}
\name{eko_produk_domestik_bruto}
\alias{eko_produk_domestik_bruto}
\title{Produk Domestik Bruto (PDB) Indonesia berdasarkan Pengeluaran dan Harga Berlaku}
\format{Dataframe dengan 625 baris dan 4 kolom

\describe{
  \item{tahun}{Tahun}
  \item{kuartal}{Kuartal}
  \item{jenis_pengeluaran}{Jenis Pengeluaran}
  \item{nilai}{Nilai hasil dari jenis pengeluaran tersebut(berdasarkan harga berlaku - dalam Miliar Rupiah)}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/produk-domestik-bruto-pdb-indonesia-berdasarkan-pengeluaran-dan-harga-berlaku}
}
\usage{
eko_produk_domestik_bruto
}
\description{
Dataset ini berisi mengenai komponen PDB Indonesia berdasarkan pengeluaran yang penghitungannya didasarkan pada harga berlaku
}
\keyword{datasets}
