% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eng.R
\docType{data}
\name{eng_domestik_batubara}
\alias{eng_domestik_batubara}
\title{Penjualan Domestik Batu Bara}
\format{Dataframe denan 456 baris dan 3 kolom

\describe{
  \item{tipe_entitas}{Tipe Entitas}
  \item{nama_entitas}{Nama Entitas}
  \item{tahun}{Tahun}
  \item{nilai}{Nilai}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/penjualan-domestik-batu-bara}
}
\usage{
eng_domestik_batubara
}
\description{
Dataset ini berisi data penjualan domestik batu bara di Indonesia
}
\keyword{datasets}
