% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eng.R
\docType{data}
\name{eng_ekspor_mineral}
\alias{eng_ekspor_mineral}
\title{Hasil Ekspor Mineral}
\format{Dataframe dengan 440 baris dan 5 kolom

\describe{
  \item{nama_perusahaan}{Nama Perusahaan}
  \item{komoditas}{Komoditas}
  \item{satuan}{Satuan berkenaan dengan komoditas tersebut}
  \item{tahun}{Tahun}
  \item{nilai}{Nilai}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/hasil-ekspor-mineral}
}
\usage{
eng_ekspor_mineral
}
\description{
Dataset ini berisi mengenai hasil ekspor mineral dari Indonesia
}
\keyword{datasets}
