% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ids.R
\docType{data}
\name{ids_kecelakaan}
\alias{ids_kecelakaan}
\title{Data Kecelakaan Industri}
\format{Dataframe dengan 14 baris dan 9 kolom

\describe{
  \item{bencana}{Jenis bencana}
  \item{tanggal}{Tanggal terjadinya bencana}
  \item{jam}{Jam terjadinya bencana}
  \item{bujur}{Titik Bujur lokasi kejadian bencana}
  \item{lintang}{Titik Lintang lokasi kejadian bencana}
  \item{lokasi}{Nama lokasi kejadian bencana}
  \item{korban}{Korban kejadia bencana}
  \item{kerugian}{Kerugian akibat bencana}
  \item{keterangan}{Keterangan lainnya mengenai bencana}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/data-kecelakaan-industri-2}
}
\usage{
ids_kecelakaan
}
\description{
Dataset ini berisi mengenai data kecelakaan industri beserta keterangannya
}
\keyword{datasets}
