% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ids.R
\docType{data}
\name{ids_umkm}
\alias{ids_umkm}
\title{Perkembangan Usaha Mikro Kecil dan Menengah (UMKM)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 9 columns.}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/perkembangan-usaha-mikro-kecil-dan-menengah-umkm}
}
\usage{
ids_umkm
}
\description{
Dataset ini berisi mengenai kondisi dan perkembangan UMKM di Indonesia. Catatan: 1. Sumbangan PDB UMKM pada tahun 1997-2000 masih didasarkan atas harga berlaku. Sumbangan PDB UMKM pada tahun 2001-2012 didasarkan atas harga konstan tahun 2000; 2. Sumbangan PDB mencakup migas dan non migas; 3. Perhitungan sumbangan PDB UMKM pada tahun 2012 masih sangat sementara dan saat ini masih dihitung ulang; 4. Nilai Ekspor UMKM hanya didasarkan pada data di sektor pertanian (pertanian, perikanan, kelautan, peternakan, kehutanan, perkebunan), industri pengolahan, dan pertambangan/penggalian
}
\keyword{datasets}
