% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ifr.R
\docType{data}
\name{ifr_kecelakaan_transportasi}
\alias{ifr_kecelakaan_transportasi}
\title{Data Kecelakaan Transportasi}
\format{Dataframe dengan 59 baris dan 10 kolom

\describe{
  \item{bencana}{Jenis bencana}
  \item{tanggal}{Tanggal terjadinya bencana}
  \item{jam}{Jam terjadinya bencana}
  \item{zona_waktu}{Zona waktu lokasi bencana}
  \item{bujur}{Titik Bujur lokasi kejadian bencana}
  \item{lintang}{Titik Lintang lokasi kejadian bencana}
  \item{lokasi}{Nama lokasi kejadian bencana}
  \item{korban}{Korban kejadia bencana}
  \item{kerugian}{Kerugian akibat bencana }
  \item{keterangan}{Keterangan lainnya mengenai bencana}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/data-kecelakaan-transportasi-2}
}
\usage{
ifr_kecelakaan_transportasi
}
\description{
Dataset ini berisi mengenai data kecelakaan transportasi beserta keterangannya pada tahun 2011 - 2014
}
\keyword{datasets}
