% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mrt.R
\docType{data}
\name{mrt_impor}
\alias{mrt_impor}
\title{Total Nilai dan Berat Impor berdasarkan Pelabuhan per Bulan}
\format{Dataframe dengan 34884 baris dan 6 kolom

\describe{
  \item{tahun}{Tahun}
  \item{bulan}{Bulan}
  \item{kode_pelabuhan}{Kode Pelabuhan}
  \item{nama_pelabuhan}{Nama Pelabuhan}
  \item{nilai}{Nilai (dalam US Dollar)}
  \item{berat}{Berat (dalam Kilogram)}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/total-nilai-dan-berat-ekspor-berdasarkan-pelabuhan-per-bulan}
}
\usage{
mrt_impor
}
\description{
Dataset ini berisi mengenai total nilai dan berat dari impor berdasarkan pelabuhan per bulan
}
\keyword{datasets}
