% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nusandata-package.R
\docType{package}
\name{nusandata}
\alias{nusandata}
\alias{nusandata-package}
\title{nusandata}
\description{
Kumpulan Data dari Portal Satu Data Indonesia
}
\examples{
library(nusandata)
data(mrt_ekspor)
str(mrt_ekspor)

# Melihat data yang tersedia di paket nusandata
data(package = "nusandata")

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aswansyahputra/nusandata}
  \item Report bugs at \url{https://github.com/aswansyahputra/nusandata/issues}
}

}
\author{
\strong{Maintainer}: Muhammad Aswan Syahputra \email{aswansyahputra@sensolution.id}

}
\keyword{internal}
