% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pgn.R
\docType{data}
\name{pgn_kekeringan}
\alias{pgn_kekeringan}
\title{Data Bencana Kekeringan}
\format{Dataframe dengan 23 baris dan 10 kolom

\describe{
  \item{bencana}{Jenis bencana}
  \item{tanggal}{Tanggal terjadinya bencana}
  \item{jam}{Jam terjadinya bencana}
  \item{bujur}{Titik Bujur lokasi kejadian bencana}
  \item{lintang}{Titik Lintang lokasi kejadian bencana}
  \item{lokasi}{Nama lokasi kejadian bencana}
  \item{korban}{Korban kejadia bencana}
  \item{kerugian}{Kerugian akibat bencana}
  \item{keterangan}{Keterangan lainnya mengenai bencana}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/data-bencana-kekeringan-2}
}
\usage{
pgn_kekeringan
}
\description{
Dataset ini berisi mengenai data bencana kekeringan beserta keterangannya
}
\keyword{datasets}
