% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pgn.R
\docType{data}
\name{pgn_rttani}
\alias{pgn_rttani}
\title{Jumlah Rumah Tangga Usaha Tani}
\format{Dataframe dengan 14 baris dan 3 kolom

\describe{
  \item{luas_lahan}{Golongan dari luas lahan (dalam meter persegi)}
  \item{tahun}{Tahun Sensus}
  \item{jumlah}{Jumlah usaha tani dengan luas lahan tersebut}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/jumlah-rumah-tangga-usaha-tani}
}
\usage{
pgn_rttani
}
\description{
Dataset ini berisi informasi mengenai jumlah usaha rumah tangga pertanian berdasarkan luas lahan yang dikelola berdasarkan hasil sensus tani
}
\keyword{datasets}
