% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sht.R
\docType{data}
\name{sht_jamban}
\alias{sht_jamban}
\title{Status Kepemilikan Fasilitas Tempat Buang Air Besar}
\format{Dataframe denan 99 baris dan 3 kolom

\describe{
  \item{nama_provinsi}{Nama Provinsi}
  \item{status_jamban}{Status dari Jamban (milik sendiri / umum / tidak ada)}
  \item{jumlah_rumah_tangga}{Jumlah Rumah Tangga dengan Status Jamban tersebut}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/status-kepemilikan-fasilitas-tempat-buang-air-besar}
}
\usage{
sht_jamban
}
\description{
Dataset ini berisi data mengenai jumlah rumah tangga dengan status kesejahteraan 30% terendah berdasarkan status kepemilikian jamban per Provinsi
}
\keyword{datasets}
