% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sht.R
\docType{data}
\name{sht_tpa_tinja}
\alias{sht_tpa_tinja}
\title{Tempat Pembuangan Akhir Tinja Rumah Tangga Miskin}
\format{Dataframe dengan 66 baris dan 3 kolom

\describe{
  \item{nama_provinsi}{Nama Provinsi}
  \item{pembuangan_akhir_tinja}{Tipe dari Pembuangan Akhir Tinja}
  \item{jumlah_rumah_tangga}{Jumlah Rumah Tangga dengan tipe Pembuangan Akhir Tinja tersebut }
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/tempat-pembuangan-akhir-tinja-rumah-tangga-miskin}
}
\usage{
sht_tpa_tinja
}
\description{
Dataset ini berisi data mengenai jumlah rumah tangga dengan status kesejahteraan 30% terendah berdasarkan jenis tempat pembuangan akhir tinja per Provinsi
}
\keyword{datasets}
