% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tks.R
\docType{data}
\name{tks_debatcapres1_2019}
\alias{tks_debatcapres1_2019}
\title{Transkrip Debat Capres 1 2019}
\format{Dataframe dengan 329 baris dan 3 kolom

\describe{
  \item{waktu}{Waktu dalam format hh:mm}
  \item{penutur}{Penutur}
  \item{teks}{Teks ucapan}
}}
\source{
Resume Debat Capres oleh BahasaKita \url{http://debatcapres.bahasakita.co.id/}
}
\usage{
tks_debatcapres1_2019
}
\description{
Dataset ini berisikan transkrip dari Debat Capres 1 yang diselenggarakan pada 17 Januari 2019. Debat Capres 2 mempertemukan antara calon presiden-wakil presiden yaitu pasangan Joko Widodo-Maruf Amin (nomor urut 01) dan Prabowo Subianto-Sandiaga Uno (nomor urut 02).
}
\keyword{datasets}
