% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ddk.R
\docType{data}
\name{ddk_tingkat_pendidikan}
\alias{ddk_tingkat_pendidikan}
\title{Tingkat Pendidikan Anak Miskin}
\format{Dataframe dengan 462 baris dan 4 kolom

\describe{
  \item{nama_provinsi}{Nama Provinsi}
  \item{tingkat_pendidikan}{Tingkat Pendidikan (SD/MI/SMP/MT/SMA/MA/Perguruan Tinggi)}
  \item{jenis_kelamin}{Jenis Kelamin}
  \item{jumlah_individu}{Jumlah dari Anak dengan Tingkat Pendidikan dan Jenis Kelamin tersebut }
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/tingkat-pendidikan-anak-miskin}
}
\usage{
ddk_tingkat_pendidikan
}
\description{
Dataset ini berisi data mengenai jumlah anak dengan status kesejahteraan 30% yang bersekolah berdasarkan tingkat pendidikan dan jenis kelamin
}
\keyword{datasets}
