% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ids.R
\docType{data}
\name{ids_umkm}
\alias{ids_umkm}
\title{Perkembangan Usaha Mikro Kecil dan Menengah (UMKM)}
\format{Dataframe dengan 16 baris dan 9 kolom

\describe{
  \item{tahun}{Tahun}
  \item{jumlah}{Jumlah UMKM}
  \item{pertumbuhan_jumla}{Pertumbuhan jumlah UMKM (persen)}
  \item{jumlah_tenaga_kerja}{Jumlah tenaga kerja UMKM (orang)}
  \item{pertumbuhan_jumlah_tenaga_kerja}{Pertumbuhan jumlah tenaga kerja UMKM (persen)}
  \item{sumbangan_pdb}{Sumbangan PDB UMKM (harga konstan, Rp. Miliar)}
  \item{sumbangan_pdb}{Pertumbuhan sumbangan PDB UMKM (persen)}
  \item{nilai_ekspor}{Nilai ekspor UMKM (Rp. Miliar)}
  \item{pertumbuhan_nilai_ekspor}{Pertumbuhan nilai ekspor UMKM (persen)}
}}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/perkembangan-usaha-mikro-kecil-dan-menengah-umkm}
}
\usage{
ids_umkm
}
\description{
Dataset ini berisi mengenai kondisi dan perkembangan UMKM di Indonesia. Catatan: 1. Sumbangan PDB UMKM pada tahun 1997-2000 masih didasarkan atas harga berlaku. Sumbangan PDB UMKM pada tahun 2001-2012 didasarkan atas harga konstan tahun 2000; 2. Sumbangan PDB mencakup migas dan non migas; 3. Perhitungan sumbangan PDB UMKM pada tahun 2012 masih sangat sementara dan saat ini masih dihitung ulang; 4. Nilai Ekspor UMKM hanya didasarkan pada data di sektor pertanian (pertanian, perikanan, kelautan, peternakan, kehutanan, perkebunan), industri pengolahan, dan pertambangan/penggalian
}
\keyword{datasets}
