% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sht.R
\docType{data}
\name{sht_tpa_tinja}
\alias{sht_tpa_tinja}
\title{Tempat Pembuangan Akhir Tinja Rumah Tangga Miskin}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 66 rows and 3 columns.}
\source{
Portal Satu Data Indonesia \url{https://data.go.id/dataset/tempat-pembuangan-akhir-tinja-rumah-tangga-miskin}
}
\usage{
sht_tpa_tinja
}
\description{
Dataset ini berisi data mengenai jumlah rumah tangga dengan status kesejahteraan 30% terendah berdasarkan jenis tempat pembuangan akhir tinja per Provinsi
}
\keyword{datasets}
